calculation-api/pyproject.toml
[tool.poetry]
name = "calculation-api"
version = "0.1.0"
description = "Different ways of calculating"
authors = ["radulovicd"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.4"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
